+++ src/PSRule/Host/ILanguageBlock.cs
@@ -3,5 +3,7 @@
     public interface ILanguageBlock
     {
         string SourcePath { get; }
+
+        string Module { get; }
     }
 }