+++ tests/PSRule.Tests/PSRule.Hint.Tests.ps1
@@ -28,11 +28,13 @@ Describe 'PSRule -- Hint keyword' -Tag 'Hint' {
 
         It 'Sets result properties' {
             $option = @{ 'Execution.InconclusiveWarning' = $False };
-            $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Option $option -Name 'HintTest' -Outcome All;
+            $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Option $option -Name 'HintTest' -Outcome All -WarningVariable outWarning -WarningAction SilentlyContinue;
+            $warningMessages = @($outWarning);
             $result | Should -Not -BeNullOrEmpty;
             $result.RuleName | Should -Be 'HintTest';
-            $result.TargetName | Should -Be 'HintTarget';
-            $result.Message | SHould -Be 'This is a message';
+            $result.Message | Should -Be 'This is a message';
+            $warningMessages.Length | Should -Be 1;
+            $warningMessages | Should -Be "Hint parameter -TargetName is obsolete and has been replaced with TargetName binding. See about_PSRule_Options for details on how to set Binding.TargetName.";
         }
 
         It 'Uses description' {
@@ -40,7 +42,7 @@ Describe 'PSRule -- Hint keyword' -Tag 'Hint' {
             $result = $testObject | Invoke-PSRule -Path (Join-Path -Path $here -ChildPath 'FromFile.Rule.ps1') -Option $option -Name 'HintTestWithDescription' -Outcome All;
             $result | Should -Not -BeNullOrEmpty;
             $result.RuleName | Should -Be 'HintTestWithDescription';
-            $result.Message | SHould -Be 'Test for Hint keyword';
+            $result.Message | Should -Be 'Test for Hint keyword';
         }
     }
 }